+++ src/main/java/ru/skypro/homework/configuration/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework;
+package ru.skypro.homework.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
