+++ src/main/java/ru/skypro/homework/controller/AuthController.java
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import ru.skypro.homework.dto.LoginReq;
-import ru.skypro.homework.dto.RegisterReq;
-import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.model.dto.LoginReq;
+import ru.skypro.homework.model.dto.RegisterReq;
+import ru.skypro.homework.model.dto.Role;
 import ru.skypro.homework.service.AuthService;
 
-import static ru.skypro.homework.dto.Role.USER;
+import static ru.skypro.homework.model.dto.Role.USER;
 
 @Slf4j
 @CrossOrigin(value = "http://localhost:3000")
