+++ src/main/java/ru/skypro/homework/model/dto/Role.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.dto;
+package ru.skypro.homework.model.dto;
 
 public enum Role {
     USER, ADMIN
