+++ src/main/java/ru/skypro/homework/model/entity/Ads.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.Model;
+package ru.skypro.homework.model.entity;
 
 import lombok.*;
 import org.hibernate.annotations.Type;
