+++ src/main/java/ru/skypro/homework/repository/AdsRepository.java
@@ -2,7 +2,7 @@ package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.skypro.homework.Model.Ads;
+import ru.skypro.homework.model.entity.Ads;
 
 @Repository
 public interface AdsRepository extends JpaRepository<Ads, Long> {
