+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -2,7 +2,7 @@ package ru.skypro.homework.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.skypro.homework.Model.User;
+import ru.skypro.homework.model.entity.User;
 
 import java.util.Optional;
