+++ src/main/java/ru/skypro/homework/service/AuthService.java
@@ -1,7 +1,7 @@
 package ru.skypro.homework.service;
 
-import ru.skypro.homework.dto.RegisterReq;
-import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.model.dto.RegisterReq;
+import ru.skypro.homework.model.dto.Role;
 
 public interface AuthService {
     boolean login(String userName, String password);
