+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -1,7 +1,7 @@
 package ru.skypro.homework.service;
 
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.Model.Image;
+import ru.skypro.homework.model.entity.Image;
 
 import java.io.IOException;
