+++ src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java
@@ -1,15 +1,13 @@
 package ru.skypro.homework.service.impl;
 
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.RegisterReq;
-import ru.skypro.homework.dto.Role;
-import ru.skypro.homework.mapper.UserMapper;
-import ru.skypro.homework.mapper.impl.RegisterReqToUserMapper;
+import ru.skypro.homework.model.dto.RegisterReq;
+import ru.skypro.homework.model.dto.Role;
+import ru.skypro.homework.model.mapper.UserMapper;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.AuthService;
 import ru.skypro.homework.service.UserService;
